requirements.txt
google-auth==2.13.0
google-auth-oauthlib==0.4.6
google-api-python-client==2.76.0
requests==2.28.2
PyPDF2==3.0.0
transformers==4.30.0
torch==2.1.0
